feat: get volume from pulseaudio sink changes
ideally this should be done by directly calling any audio interface
available on the system, like pipewire, pulseaudio or alsa. Pipewire
should give the least latency. But libraries and bindings are a little
bit obscure.

With this method, we run a subprocess in a thread, subscribe to it and
parse its volume accordingly